Obviously has to be non-empty, non-whitespace
To stop the tactile theme from having jekyll-titles-from-headings using the "page header" (ie the first line with a header mark) as the page title and thus overwrite the site title in the header section: put a single non-whitespace character on the first line. Will look ugly, but as the tactile theme and jekyll-titles-from-headings have the same author, I don't think it's easy to make it do my bidding otherwise.